Extract `SyntaxFormatter`.
